Restructured operations code for better reading.
